Added ClassClass Candidate finder
